backend/requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow
scikit-learn
joblib
python-multipart
